+++ Modules/UserManagement/Http/Requests/StoreNewUserPassword.php
@@ -34,7 +34,7 @@ class StoreNewUserPassword extends FormRequest
             'password' => [
                 'required',
                 'string',
-                'min:6',
+                'min:8',
                 'pwned',
                 'confirmed',
             ],