+++ Modules/UserManagement/Http/Requests/StoreUser.php
@@ -42,7 +42,7 @@ class StoreUser extends FormRequest
             'password' => [
                 'required',
                 'string',
-                'min:6',
+                'min:8',
                 'pwned',
             ],
             'roles' => [