+++ Modules/UserManagement/Http/Requests/UpdateUser.php
@@ -42,7 +42,7 @@ class UpdateUser extends FormRequest
             'password' => [
                 'nullable',
                 'string',
-                'min:6',
+                'min:8',
                 'pwned',
             ],
             'roles' => [